pyproject.toml
[project]
name = "coda-teammate-suite"
version = "0.1.0"
description = "AI teammate orchestration suite"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langgraph-checkpoint",
    "langgraph-checkpoint-postgres",
    "psutil>=7.2.2",
    "psycopg-binary",
    "pydantic-settings",
    "pytest-asyncio",
    "ruff",
    "langchain-deepseek>=1.0.1",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "litellm>=1.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "fastmcp>=1.0.0",
    "python-fasthtml>=0.6.0",
    "fastapi>=0.128.0",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
    "apps/liaison_brain",
    "apps/workspace_ui",
    "packages/skillhive",
    "packages/sentry_foundation",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
